Move define_common_arguments to cli.py
