+++ temboardagent/cli.py
@@ -11,6 +11,15 @@ from .logger import LOG_FORMAT
 logger = logging.getLogger(__name__)
 
 
+def define_common_arguments(parser):
+    parser.add_argument(
+        '-c', '--config',
+        action='store', dest='configfile',
+        default='/etc/temboard-agent/temboard-agent.conf',
+        help="Configuration file. Default: %(default)s",
+    )
+
+
 def cli(main):
     # A decorator to add consistent CLI bootstrap
     #
